revert handling
