+++ packages/localnet/src/evmDepositAndCall.ts
@@ -1,4 +1,3 @@
-import * as UniswapV2Router02 from "@uniswap/v2-periphery/build/UniswapV2Router02.json";
 import * as ZRC20 from "@zetachain/protocol-contracts/abi/ZRC20.sol/ZRC20.json";
 import { ethers } from "ethers";
 
@@ -6,6 +5,7 @@ import { evmOnRevert } from "./evmOnRevert";
 import { log, logErr } from "./log";
 import { zetachainDepositAndCall } from "./zetachainDepositAndCall";
 import { zetachainOnAbort } from "./zetachainOnAbort";
+import { zetachainSwapToCoverGas } from "./zetachainSwapToCoverGas";
 
 export const evmDepositAndCall = async ({
   tss,
@@ -52,9 +52,8 @@ export const evmDepositAndCall = async ({
     return;
   }
 
-  const zrc20 = foreignCoin.zrc20_contract_address;
   try {
-    zetachainDepositAndCall({
+    await zetachainDepositAndCall({
       args,
       chainID,
       foreignCoins,
@@ -68,33 +67,21 @@ export const evmDepositAndCall = async ({
     }
     logErr("ZetaChain", `onCall failed: ${err}`);
     const revertOptions = args[5];
-    const zrc20Contract = new ethers.Contract(zrc20, ZRC20.abi, deployer);
-    const [gasZRC20, gasFee] = await zrc20Contract.withdrawGasFeeWithGasLimit(
-      revertOptions[4]
-    );
-    let revertAmount;
-    let revertGasFee = gasFee;
-    let isGas = true;
-    let token = null;
-    if (zrc20 !== gasZRC20) {
-      token = foreignCoins.find(
-        (coin) => coin.zrc20_contract_address === zrc20
-      )?.asset;
-      isGas = false;
-      revertGasFee = await swapToCoverGas(
-        deployer,
-        zrc20,
-        gasZRC20,
-        gasFee,
+    const gasLimit = revertOptions[4];
+    const { revertGasFee, isGas, token, zrc20 } = await zetachainSwapToCoverGas(
+      {
+        foreignCoins,
         amount,
-        await fungibleModuleSigner.getAddress(),
-        zrc20Contract,
+        asset,
+        chainID,
+        deployer,
+        fungibleModuleSigner,
         provider,
-        protocolContracts.wzeta.target,
-        protocolContracts.uniswapRouterInstance.target
-      );
-    }
-    revertAmount = amount - revertGasFee;
+        protocolContracts,
+        gasLimit,
+      }
+    );
+    const revertAmount = amount - revertGasFee;
     if (revertAmount > 0) {
       return await evmOnRevert({
         amount: revertAmount,
@@ -120,6 +107,7 @@ export const evmDepositAndCall = async ({
       const revertMessage = revertOptions[3];
       log("ZetaChain", `Transferring tokens to abortAddress ${abortAddress}`);
       deployer.reset();
+      const zrc20Contract = new ethers.Contract(zrc20, ZRC20.abi, deployer);
       const transferTx = await zrc20Contract.transfer(abortAddress, amount);
       await transferTx.wait();
       return await zetachainOnAbort({
@@ -135,105 +123,4 @@ export const evmDepositAndCall = async ({
       });
     }
   }
-};
-
-const swapToCoverGas = async (
-  deployer: any,
-  zrc20: string,
-  gasZRC20: string,
-  gasFee: any,
-  amount: any,
-  fungibleModule: any,
-  zrc20Contract: any,
-  provider: any,
-  wzeta: string,
-  router: string
-) => {
-  /**
-   * Retrieves the amounts for swapping tokens using UniswapV2.
-   * @param {"in" | "out"} direction - The direction of the swap ("in" or "out").
-   * @param {any} provider - The ethers provider.
-   * @param {any} amount - The amount to swap.
-   * @param {string} tokenA - The address of token A.
-   * @param {string} tokenB - The address of token B.
-   * @returns {Promise<any>} - The amounts for the swap.
-   * @throws Will throw an error if the UniswapV2 router address cannot be retrieved.
-   */
-  const getAmounts = async (
-    direction: "in" | "out",
-    provider: any,
-    amount: any,
-    tokenA: string,
-    tokenB: string,
-    routerAddress: any,
-    routerABI: any
-  ) => {
-    if (!routerAddress) {
-      throw new Error("Cannot get uniswapV2Router02 address");
-    }
-
-    const uniswapRouter = new ethers.Contract(
-      routerAddress,
-      routerABI.abi,
-      provider
-    );
-
-    const path = [tokenA, tokenB];
-
-    const amounts =
-      direction === "in"
-        ? await uniswapRouter.getAmountsIn(amount, path)
-        : await uniswapRouter.getAmountsOut(amount, path);
-    return amounts;
-  };
-
-  const uniswapV2Router = new ethers.Contract(
-    router,
-    UniswapV2Router02.abi,
-    deployer
-  );
-  deployer.reset();
-  const approvalTx = await zrc20Contract.approve(router, amount);
-  await approvalTx.wait();
-
-  const path = [zrc20, wzeta, gasZRC20];
-
-  const deadline = Math.floor(Date.now() / 1000) + 60 * 20;
-  const maxZRC20ToSpend = amount;
-
-  try {
-    const swapTx = await uniswapV2Router.swapTokensForExactTokens(
-      gasFee,
-      maxZRC20ToSpend,
-      path,
-      fungibleModule,
-      deadline
-    );
-
-    await swapTx.wait();
-  } catch (swapError) {
-    logErr("ZetaChain", `Error performing swap on Uniswap: ${swapError}`);
-  }
-
-  const amountInZeta = await getAmounts(
-    "in",
-    provider,
-    gasFee,
-    wzeta,
-    gasZRC20,
-    router,
-    UniswapV2Router02
-  );
-
-  const amountInZRC20 = await getAmounts(
-    "in",
-    provider,
-    amountInZeta[0],
-    zrc20,
-    wzeta,
-    router,
-    UniswapV2Router02
-  );
-
-  return amountInZRC20[0];
 };
